requirements.txt
streamlit
langchain-community
openai
chromadb
tiktoken
pymupdf
python-dotenv
faiss-cpu